+++ fluent_pages/tests/menu.py
@@ -21,7 +21,7 @@ class MenuTests(AppTestCase):
         level1b = SimpleTextPage.objects.create(title="Level1b", slug="level1b", parent=root, status=SimpleTextPage.PUBLISHED, author=cls.user)
 
 
-    def test_menu(self):
+    def test_navigation(self):
         """
         The API should return the top level
         """
@@ -46,7 +46,7 @@ class MenuTests(AppTestCase):
         # NOTE: does not support sub pages.
 
 
-    def test_active_menu_item(self):
+    def test_menu_items(self):
         """
         The menu API should return the active item.
         """
@@ -54,13 +54,21 @@ class MenuTests(AppTestCase):
 
         nav = Page.objects.toplevel_navigation(current_page=current_page)
         menu = [PageNavigationNode(page, current_page=current_page) for page in nav]
+
+        # Test structure
         self.assertEqual(menu[0].slug, 'home')
         self.assertEqual(menu[1].slug, 'root2')
+
+        # PageNavigationNode.parent should deal with missing get_parent() attribute:
+        self.assertEqual(menu[0].parent, None)
+        self.assertEqual(menu[1].parent, None)
+
+        # Test active state
         self.assertEqual(menu[0].is_active, False)
         self.assertEqual(menu[1].is_active, True)
 
 
-    def test_active_sub_menu_item(self):
+    def test_sub_menu_items(self):
         """
         The menu API should return the active item.
         """
@@ -68,13 +76,21 @@ class MenuTests(AppTestCase):
 
         nav = Page.objects.toplevel_navigation(current_page=current_page)
         menu = [PageNavigationNode(page, current_page=current_page) for page in nav]
+
+        # Test structure
         self.assertEqual(menu[0].slug, 'home')
         self.assertEqual(menu[1].slug, 'root2')
-        self.assertEqual(menu[0].is_active, False)
-        self.assertEqual(menu[1].is_active, False)
 
         children = list(menu[0].children)
         self.assertEqual(children[0].slug, 'level1a')
         self.assertEqual(children[1].slug, 'level1b')
+
+        # Test reverse structure
+        self.assertEqual(children[0].parent, menu[0])
+
+        # Test active states
+        self.assertEqual(menu[0].is_active, False)
+        self.assertEqual(menu[1].is_active, False)
+
         self.assertEqual(children[0].is_active, True)
         self.assertEqual(children[1].is_active, False)